+++ .gitignore
@@ -27,4 +27,3 @@ container.jpg
 libglad.a
 lib
 libmap_editor_lib.a
-textures
